Added src folder, refactored saving, loading and classes
