+++ app.js
@@ -2,6 +2,10 @@
 const fs = require('fs');
 const express = require('express');
 const ejs = require('ejs');
+const BlogPost = require('./src/BlogPost.js');
+const Comment = require('./src/Comment.js');
+const File = require('./src/fileHandler.js');
+const jsonFilePath = 'public/assets/blogs.json';
 //setup for express
 const app = express();
 const port = 3000;
@@ -11,76 +15,13 @@ app.use(express.static('views'));
 app.use(express.urlencoded());
 
 
-class Comment {
-    constructor(content, author) {
-        this.content = content;
-        this.time = Date(Date.now).toString();
-        if (author)
-            this.author = author;
-        else
-            this.author = "anonymous";
-    }
-}
-class BlogPost {
-    constructor(title, content, author) {
-        this.id = 1;
-        this.title = title;
-        this.content = content;
-        if (author)
-            this.author = author;
-        else
-            this.author = "anonymous";
-        this.comments = [];
-        this.reactions = [0, 0, 0];
-    }
-
-    addComment(comment, author) {
-        this.comments.push(new Comment(comment, author));
-    }
-}
-
-
-//saves new json object to array
-const saveJson = (jsonObj) => {
-    fs.readFile('public/assets/blogs.json', (err, data) => {
-        if (err) throw err;
-        let jsonArray;
-        try {
-            jsonArray = JSON.parse(data);
-        }
-        catch (err) {
-            console.log(err);
-            //unexpected end of json input due to empty json file
-            jsonArray = [];
-        }
-        jsonObj.id = jsonArray.length + 1;
-        jsonArray.push(jsonObj);
-        //convert array to form that can be saved in json file
-        newJsonArray = JSON.stringify(jsonArray);
-        fs.writeFile("public/assets/blogs.json", newJsonArray, () => {
-        });
-        
-    });
-    
-    
-}
 
 //first page render
 app.get('', (req, res) => {
-    fs.readFile('public/assets/blogs.json', (err, data) => {
-        let jsonArray;
-        try {
-            jsonArray = JSON.parse(data);
-        }
-        catch (err) {
-            console.log(err);
-            //unexpected end of json input due to empty json file
-            jsonArray = [];
-        
-        }
-        res.render('homepage.html',{blogs : jsonArray });
-   
-});
+    File.readFile(jsonFilePath).then(response =>
+    {
+         res.render('homepage.html',{blogs : response });
+    });
 });
 
 //blog creation render
@@ -92,18 +33,14 @@ app.get('/create', (req,res) => {
 //gets and renders a specific blog post
 app.get('/blog/:index', (req, res) => {
     let index = req.params.index;
-    fs.readFile('public/assets/blogs.json', (err, data) => {
-        if (err) throw err;
-        //parses data from json file into a useable array format
-        let jsonArray = JSON.parse(data);
-        //checks input index is a valid one
-        if (index > 0 && index <= jsonArray.length) {
-            res.render("blogPage.html", jsonArray[index - 1]);
+    File.readFile(jsonFilePath).then(response => {
+        if (index > 0 && index <= response.length) {
+            res.render("blogPage.html", response[index - 1]);
         }
         else {
-            res.send(`Blog doesn't exist. Please choose an index between 1 and ${jsonArray.length}`);
+            res.send(`Blog doesn't exist. Please choose an index between 1 and ${response.length}`);
         }
-    });
+    })  
 });
 
 //when form data is submitted
@@ -111,15 +48,18 @@ app.post('/submit', (req, res) => {
     let title = req.body.title;
     let content = req.body.content;
     let author = req.body.author;
-    let newPost = new BlogPost(
+    //error handling for empty blog posts
+    if(title !== "" && content !== ""){
+        let newPost = new BlogPost(
         title,
         content,
         author
     );
-   saveJson(newPost);
-   
+   File.saveJson(newPost);
+   res.redirect(`/`);
+    }
     
-    res.redirect(`/`);
+    
 });
 
 app.post('/comment', (req, res) => {
@@ -129,21 +69,16 @@ app.post('/comment', (req, res) => {
     let content = req.body.comment;
     let author = req.body.author;
     //if the comment is not empty
-    if(content !== "")
-    {
     // get the index from the url of the post request
     let index = reqUrl.split('/')[4];
-
+    if(content !== "")
+    {
     //get our array of blogs
-    fs.readFile('public/assets/blogs.json', (err, data) => {
-        if (err) throw err;
-        //parses data from json file into a useable array format
-        let jsonArray = JSON.parse(data);
-        //add comment to specific blog by id
+    File.readFile(jsonFilePath).then(response => {
         comment = new Comment(content,author);
-        jsonArray[index -1].comments.push(comment);
+        response[index -1].comments.push(comment);
 
-        newJsonArray = JSON.stringify(jsonArray);
+        newJsonArray = JSON.stringify(response);
         //save the blogs back to file
         fs.writeFile("public/assets/blogs.json", newJsonArray, () => {});
