+++ public/assets/blogs.json
@@ -1,1 +1,1 @@
-[{"id":1,"title":"I work with a bunch of cheats","content":":(","author":"Lauren","comments":[{"content":"i  can't think","time":"Tue Nov 26 2019 15:51:36 GMT+0000 (Greenwich Mean Time)","author":"josh"}],"reactions":[0,0,0]},{"id":2,"title":"wekjfj","content":"sd  fsd fnsf","author":"anonymous","comments":[{"content":":)","time":"Tue Nov 26 2019 15:30:13 GMT+0000 (Greenwich Mean Time)","author":"anonymous"},{"content":"hiiiiiiiiiiiiiii","time":"Tue Nov 26 2019 15:31:34 GMT+0000 (Greenwich Mean Time)","author":"lauren"}],"reactions":[0,0,0]},{"id":3,"title":"I love Lauren","content":"AKhaHFSK","author":"Levi","comments":[{"content":"get away from my man","time":"Tue Nov 26 2019 16:21:15 GMT+0000 (Greenwich Mean Time)","author":"JOSH"}],"reactions":[0,0,0]},{"id":4,"title":"Hello darkness, my old friend","content":"kajdkajsdkajsbdkajbdjkasd","author":"Lozza","comments":[],"reactions":[0,0,0]},{"id":5,"title":"dfgdfg","content":"dfgdfgdfg","author":"dgdfg","comments":[],"reactions":[0,0,0]}]
+[{"id":1,"title":"I work with a bunch of cheats","content":":(","author":"Lauren","comments":[{"content":"i  can't think","time":"Tue Nov 26 2019 15:51:36 GMT+0000 (Greenwich Mean Time)","author":"josh"}],"reactions":[0,0,0]},{"id":2,"title":"wekjfj","content":"sd  fsd fnsf","author":"anonymous","comments":[{"content":":)","time":"Tue Nov 26 2019 15:30:13 GMT+0000 (Greenwich Mean Time)","author":"anonymous"},{"content":"hiiiiiiiiiiiiiii","time":"Tue Nov 26 2019 15:31:34 GMT+0000 (Greenwich Mean Time)","author":"lauren"}],"reactions":[0,0,0]},{"id":3,"title":"I love Lauren","content":"AKhaHFSK","author":"Levi","comments":[{"content":"get away from my man","time":"Tue Nov 26 2019 16:21:15 GMT+0000 (Greenwich Mean Time)","author":"JOSH"}],"reactions":[0,0,0]},{"id":4,"title":"Hello darkness, my old friend","content":"kajdkajsdkajsbdkajbdjkasd","author":"Lozza","comments":[],"reactions":[0,0,0]},{"id":5,"title":"dfgdfg","content":"dfgdfgdfg","author":"dgdfg","comments":[],"reactions":[0,0,0]},{"id":6,"title":"2","content":"3","author":"1","comments":[],"reactions":[0,0,0]},{"id":7,"title":"23","content":"23","author":"32","comments":[],"reactions":[0,0,0]},{"id":8,"title":"2","content":"2","author":"anonymous","comments":[],"reactions":[0,0,0]}]
