+++ public/js/blogPage.js
@@ -1,7 +1,4 @@
-import {saveFile, area} from "./fileHandler.js";
 $(document).ready (() => {
-    console.log(saveFile());
-    console.log(area(2));
     let maxLength = 300;
     $('#comment').keyup( ()=> {
         let textLength = maxLength - $('#comment').val().length;
