--- public/js/fileHandler.js
@@ -1,9 +1,0 @@
-const PI = 3.14159265359;
-
-export function area(radius) {
-  return (radius ** 2) * PI;
-}
-
-export  function saveFile() {
-  return 3;
-}
